+++ src/compas_rhino/geometry/brep/trim.py
@@ -7,9 +7,11 @@ import Rhino  # type: ignore
 from compas.geometry import BrepTrim
 from compas_rhino.geometry import RhinoNurbsCurve
 
+from .vertex import RhinoBrepVertex
+
 
 class RhinoBrepTrim(BrepTrim):
-    """An interface for a Brep Trim
+    """Trims hold topological information about how the edges of a Brep face are are organized.
 
     Attributes
     ----------
@@ -21,6 +23,12 @@ class RhinoBrepTrim(BrepTrim):
         True if this trim is reversed from its associated edge curve and False otherwise.
     native_trim : :class:`Rhino.Geometry.BrepTrim`
         The underlying Rhino BrepTrim object.
+    start_vertex : :class:`compas_rhino.geometry.RhinoBrepVertex`, read-only
+        The start vertex of this trim.
+    end_vertex : :class:`compas_rhino.geometry.RhinoBrepVertex`, read-only
+        The end vertex of this trim.
+    vertices : list[:class:`compas_rhino.geometry.RhinoBrepVertex`], read-only
+        The list of vertices which comprise this trim (start and end).
 
     """
 
@@ -30,6 +38,8 @@ class RhinoBrepTrim(BrepTrim):
         self._curve = None
         self._is_reversed = None
         self._iso_type = None
+        self._start_vertex = None
+        self._end_vertex = None
         if rhino_trim:
             self.native_trim = rhino_trim
 
@@ -40,7 +50,8 @@ class RhinoBrepTrim(BrepTrim):
     @property
     def __data__(self):
         return {
-            "vertex": self._trim.StartVertex.VertexIndex,
+            "start_vertex": self._trim.StartVertex.VertexIndex,
+            "end_vertex": self._trim.EndVertex.VertexIndex,
             "edge": self._trim.Edge.EdgeIndex if self._trim.Edge else -1,  # singular trims have no associated edge
             "curve": RhinoNurbsCurve.from_rhino(self._trim.TrimCurve.ToNurbsCurve()).__data__,
             "iso": str(self._trim.IsoStatus),
@@ -68,7 +79,7 @@ class RhinoBrepTrim(BrepTrim):
         curve = RhinoNurbsCurve.__from_data__(data["curve"]).rhino_curve
         iso_status = getattr(Rhino.Geometry.IsoStatus, data["iso"])
         is_reversed = True if data["is_reversed"] == "true" else False
-        instance.native_trim = builder.add_trim(curve, data["edge"], is_reversed, iso_status, data["vertex"])
+        instance.native_trim = builder.add_trim(curve, data["edge"], is_reversed, iso_status, data["start_vertex"])
         return instance
 
     # ==============================================================================
@@ -78,6 +89,18 @@ class RhinoBrepTrim(BrepTrim):
     @property
     def curve(self):
         return self._curve
+
+    @property
+    def start_vertex(self):
+        return self._start_vertex
+
+    @property
+    def end_vertex(self):
+        return self._end_vertex
+
+    @property
+    def vertices(self):
+        return [self.start_vertex, self.end_vertex]
 
     @property
     def is_reverse(self):
@@ -97,3 +120,5 @@ class RhinoBrepTrim(BrepTrim):
         self._curve = RhinoNurbsCurve.from_rhino(rhino_trim.TrimCurve.ToNurbsCurve())
         self._is_reversed = rhino_trim.IsReversed()
         self._iso_type = int(rhino_trim.IsoStatus)
+        self._start_vertex = RhinoBrepVertex(rhino_trim.StartVertex)
+        self._end_vertex = RhinoBrepVertex(rhino_trim.EndVertex)